From Kristofer Tingdahl, warning fixes.
git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@15030 16af8721-9629-0410-8352-f15c8da7e697

+++ src/osgViewer/ViewerEventHandlers.cpp
@@ -721,8 +721,8 @@ bool InteractiveImageHandler::mousePosition(osgViewer::View* view, osg::NodeVisi
     if (!view) return false;
     if (_fullscreen)
     {
-        x = ea.getX();
-        y = ea.getY();
+        x = (int) ea.getX();
+        y = (int) ea.getY();
         return true;
     }
